fix: address all 5 Greptile review issues
1. batch_get_file_contents: preserve file path on rejection via
   per-file try/catch in shared batchGetFiles()
2. configure "set": only "debug" is truly immediate — move timeout,
   verifyWrites, maxResponseChars to restart-required settings
3. handleRecentChanges: batch API calls (size 20) instead of
   unbounded Promise.allSettled on all vault files
4. read-only preset: remove open_file from consolidated read-only
   (POST /open can create files)
5. Extract shared helpers into src/tools/shared.ts — eliminates
   duplication of 14 functions across granular.ts and consolidated.ts
